switch miner to an interface
